+++ plugin-build/plugin/src/main/java/com/ncorti/ktfmt/gradle/KtfmtPlugin.kt
@@ -53,10 +53,7 @@ abstract class KtfmtPlugin : Plugin<Project> {
         project.tasks.withType(KtfmtBaseTask::class.java).configureEach {
             it.ktfmtClasspath.from(ktFmt)
             it.formattingOptionsBean.set(ktfmtExtension.toFormattingOptions())
-            it.useClassloaderIsolation.set(
-                project.findProperty("ktfmt.useClassloaderIsolation")?.toString()?.toBoolean() ==
-                    true
-            )
+            it.useClassloaderIsolation.set(ktfmtExtension.useClassloaderIsolation)
         }
 
         topLevelFormat = createTopLevelFormatTask(project)